User click on cell updates game object in store.js, but API request fails
This code updates a single value in the cells array of the game object.
But upon pushing that data through in an api request, I get 400 Bad Request.

Adding this as a pointer for the following issue:

https://github.com/ga-wdi-boston/game-project/issues/753

Ideas: I have not yet joined player O. Is it possible I am getting a bad
request because it doesn't have data for the player O player?